Fix failing test
The test was expecting to divide the buffer into three PDUs, but now that the minimum MAX_PDU has been reduced to 1K instead of 4K, it is being divided in quite some more fragments. Reducing the buffer from 9K to 2.5K makes the buffer be divided into three PDUs again.